backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
tritonclient[grpc]==2.41.0
prometheus-fastapi-instrumentator==6.1.0
pydantic==2.5.0
pydantic-settings==2.1.0

